Factor out Option
